Pass ordering to defer_iteration and shard by offest only.
* If we have one inequality filter in queryset we are passing to
defer_iteration, sharding did not work, as we would end up with two
inequality filters. To fix the issue, instead of adding new inequality
filter (`filter(pk__gt=offset_pk)`) we simply slice the queryset with
given offset: `[offset:offset + shard_size]`.
* we order the queryset by the any filter provided in `order_by`
parameter (which is needed if we use an inequality filter in the given
queryset) and by `pk`.